Tested logic to kill the signup link after use

+++ oldmail/views.py
@@ -217,14 +217,16 @@ class ProfileVerifyView(DetailView):
     def get_object(self, **kwargs):
         signup_link = get_object_or_404(SignupLink, random_string=self.kwargs['random_string'], account__slug=self.kwargs['slug'])
         profile = get_object_or_404(Profile, user__email=signup_link.email)
-        profile.is_verified = True
-        profile.save()
+        if profile.is_verified:
+            signup_link = get_object_or_404(SignupLink, random_string=self.kwargs['random_string'], account__slug=self.kwargs['slug'])
+            signup_link.delete()
+        else:
+            profile.is_verified = True
+            profile.save()
 
         return profile
 
     def render_to_response(self, context):
-        signup_link = get_object_or_404(SignupLink, random_string=self.kwargs['random_string'], account__slug=self.kwargs['slug'])
-        signup_link.delete()
         return HttpResponseRedirect(reverse('account_invite', args=[self.get_object().account.slug]))
